perf(sentry): inline config into instrumentation-client + nav hook
Two related Turbopack-future-proofing changes the Next 16 build
explicitly asked for:

1. Sentry deprecation: `sentry.client.config.ts` will stop working
   under Turbopack. Move all of its contents (Sentry.init + the
   lazyLoadIntegration replay deferral added in commit 5) into
   `instrumentation-client.ts`. Delete the old standalone file. The
   prior `instrumentation-client.ts` was a one-line re-import of the
   old file, so this is functionally identical at runtime.

2. Add the missing `onRouterTransitionStart` export pointing at
   `Sentry.captureRouterTransitionStart`. Required by @sentry/nextjs
   in Next.js App Router for client-side navigation traces. The build
   was logging "ACTION REQUIRED" on every run; Sentry's nav waterfalls
   light up after this.

No new bundle weight — just consolidates two files into one and exports
a hook the SDK was already looking for.
